QR scanner set to zxing for foss version
